requirements.txt
# Core dependencies
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
diskcache==5.6.3
filelock==3.20.0
fsspec==2025.12.0
idna==3.11
jinja2==3.1.6

# LLM
llama-cpp-python==0.3.16

# Audio processing
llvmlite==0.46.0
numba==0.63.1
numpy==2.3.5
scipy==1.17.0
sounddevice==0.5.3
soundfile==0.13.1
pyaudio==0.2.14

# Speech recognition
openai-whisper==20250625
tiktoken==0.12.0

# Text-to-Speech (improved with Kokoro)
pyttsx3==2.99
kokoro-tts  # Natural human-like voice

# PyTorch (CPU version)
torch==2.9.1+cpu
torchaudio==2.9.1+cpu

# Utilities
markupsafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
pycparser==2.23
pyyaml==6.0.2  # For config file
regex==2026.1.15
requests==2.32.5
setuptools==70.2.0
sympy==1.14.0
tqdm==4.67.1
triton==3.5.1
typing-extensions==4.15.0
urllib3==2.6.3